SyntaxParser.java: Move into parser package
