+++ src/main/java/seedu/address/logic/parser/FindCommandParser.java
@@ -3,7 +3,6 @@ package seedu.address.logic.parser;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.conditionalparser.SyntaxParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
